+++ packages/tddy-web/src/components/sessions/SessionsDrawerScreen.tsx
@@ -652,7 +652,11 @@ export function SessionsDrawerScreen({
             />
           )}
         </div>
-        <HostStatsFooter attachment={attachment} />
+        <HostStatsFooter
+          attachment={attachment}
+          runtimes={runtimes}
+          runtimeRegistry={runtimeRegistry}
+        />
       </AppShell>
     </TooltipProvider>
     </UploadProgressProvider>